Create README.md
The project focuses on analysing the impact of various factors on the occurrence of war-related fires in Ukraine during the ongoing conflict that began on February 24, 2022. The project utilizes a dataset containing information on population density, urban areas, daily fire occurrences, proximity to conflict zones, and other relevant features. The objective is to build a regression model to understand the relationships between these factors and the frequency of war-related fires in Ukraine-held areas.

Through statistical analysis the project aims to identify significant predictors, assess their impact, and provide insights that can aid in resource allocation and decision- making. The analysis may contribute to a better understanding of the dynamics of fire incidents during times of conflict and provide valuable information for humanitarian and strategic planning in the affected regions.